+++ src/pharmpy/modeling/__init__.py
@@ -69,6 +69,7 @@ from pharmpy.modeling.ml import (
     predict_outliers,
 )
 from pharmpy.modeling.odes import (
+    has_mixed_mm_fo_elimination,
     add_individual_parameter,
     add_lag_time,
     add_peripheral_compartment,
@@ -193,6 +194,7 @@ __all__ = [
     'has_combined_error_model',
     'has_first_order_elimination',
     'has_michaelis_menten_elimination',
+    'has_mixed_mm_fo_elimination',
     'has_proportional_error_model',
     'has_zero_order_absorption',
     'has_zero_order_elimination',